This is an attempt to integrate ADC code into the http server example.
the http server project builds and runs, but produces a garbage voltage.  

The next step is to refine the ADC to produce an accurate wind speed
based on my instrument.

the http server example requires configuration via smartphone; another
next step is to configure the device at build-time and modify the web
page.