Getting closer
This program reads the ADC, takes an average, calculates a somewhat
accurate wind speed. and passes the value as an INTEGER to the web page.

The value is not refreshing at all. the value is passed once and does
not update automatically (as it should) or upon page refresh. This means
I am not actually polling the token, just reading it once. This could be
because my OS hits a snag, the JS functions are not seu up correctly, or
I am not updating the token how I think I am.

To do:
	eliminate the need for smartphone configuration upon reset
	implement dynamic/static wind speed updating
	implement requirement of .1 knot resolution, instead of 1kt